Modularized codebase by moving our new code out of applyGQL as much as possible
